requirements.txt
APScheduler==3.11.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-apscheduler==0.7.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.45.1
